Fix missing cstdint includes for GCC and add libwayland-bin

+++ src/formats/sox_text.h
@@ -2,13 +2,14 @@
 
 #include "formats/file_format.h"
 
+#include <cstdint>
 #include <string>
 #include <vector>
 
 namespace kuf {
 
 struct TextEntry {
-    uint8_t maxLength;
+    uint16_t maxLength;
     std::string text;
 };
